+++ apps/backend/src/modules/project-groups/libs/types/types.ts
@@ -1,6 +1,8 @@
 export {
 	type ProjectGroupCreateRequestDto,
 	type ProjectGroupCreateResponseDto,
+	type ProjectGroupGetAllRequestDto,
+	type ProjectGroupGetAllResponseDto,
 	type ProjectGroupUpdateRequestDto,
 	type ProjectGroupUpdateResponseDto,
 } from "@git-fit/shared";