+++ apps/frontend/src/pages/project-access-management/libs/helpers/get-project-group-columns.helper.tsx
@@ -1,0 +1,23 @@
+import { formatDate } from "~/libs/helpers/helpers.js";
+import { type TableColumn } from "~/libs/types/types.js";
+
+import { type ProjectGroupRow } from "../types/types.js";
+
+const getProjectGroupColumns = (): TableColumn<ProjectGroupRow>[] => [
+	{
+		accessorKey: "name",
+		header: "Name",
+	},
+	{
+		accessorFn: (projectGroup: ProjectGroupRow): string =>
+			projectGroup.permissions.join(", "),
+		header: "Permissions",
+	},
+	{
+		accessorFn: (projectGroup: ProjectGroupRow): string =>
+			formatDate(new Date(projectGroup.createdAt), "d MMM yyyy HH:mm"),
+		header: "Created At",
+	},
+];
+
+export { getProjectGroupColumns };