+++ packages/shared/src/index.ts
@@ -93,11 +93,19 @@ export {
 	type ProjectGroupCreateResponseDto,
 	projectGroupCreateValidationSchema,
 	ProjectGroupError,
+	type ProjectGroupGetAllItemResponseDto,
+	type ProjectGroupGetAllRequestDto,
+	type ProjectGroupGetAllResponseDto,
 	ProjectGroupsApiPath,
 	type ProjectGroupUpdateRequestDto,
 	type ProjectGroupUpdateResponseDto,
 	projectGroupUpdateValidationSchema,
 } from "./modules/project-groups/project-groups.js";
+export {
+	ProjectPermissionsApiPath,
+	type ProjectPermissionsGetAllItemResponseDto,
+	type ProjectPermissionsGetAllResponseDto,
+} from "./modules/project-permissions/project-permissions.js";
 export {
 	type ProjectCreateRequestDto,
 	projectCreateValidationSchema,