+++ packages/shared/src/modules/project-groups/libs/types/types.ts
@@ -1,4 +1,7 @@
 export { type ProjectGroupCreateRequestDto } from "../types/project-group-create-request-dto.type.js";
 export { type ProjectGroupCreateResponseDto } from "../types/project-group-create-response-dto.type.js";
-export { type ProjectGroupUpdateRequestDto } from "./project-group-update-request-dto.type.js";
+export { type ProjectGroupGetAllItemResponseDto } from "../types/project-group-get-all-item-response-dto.type.js";
+export { type ProjectGroupGetAllRequestDto } from "./project-group-get-all-request-dto.type.js";
+export { type ProjectGroupGetAllResponseDto } from "./project-group-get-all-response-dto.type.js";
+export { type ProjectGroupUpdateRequestDto } from "./project-group-patch-request-dto.type.js";
 export { type ProjectGroupUpdateResponseDto } from "./project-group-update-response-dto.type.js";