+++ packages/shared/src/modules/project-groups/project-groups.ts
@@ -3,6 +3,9 @@ export { ProjectGroupError } from "./libs/exceptions/exceptions.js";
 export {
 	type ProjectGroupCreateRequestDto,
 	type ProjectGroupCreateResponseDto,
+	type ProjectGroupGetAllItemResponseDto,
+	type ProjectGroupGetAllRequestDto,
+	type ProjectGroupGetAllResponseDto,
 	type ProjectGroupUpdateRequestDto,
 	type ProjectGroupUpdateResponseDto,
 } from "./libs/types/types.js";