+++ packages/shared/src/modules/project-permissions/libs/types/project-permissions-get-all-item-response-dto.type.ts
@@ -1,0 +1,7 @@
+type ProjectPermissionsGetAllItemResponseDto = {
+	id: number;
+	key: string;
+	name: string;
+};
+
+export { type ProjectPermissionsGetAllItemResponseDto };