+++ packages/shared/src/modules/users/libs/types/user-auth-response-dto.type.ts
@@ -12,6 +12,10 @@ type UserAuthResponseDto = {
 	}[];
 	id: number;
 	name: string;
+	projectGroups: {
+		id: number;
+		name: string;
+	}[];
 };
 
 export { type UserAuthResponseDto };